refactor: split client into client-h1/h2
