+++ lib/core/symbols.js
@@ -33,6 +33,8 @@ module.exports = {
   kNeedDrain: Symbol('need drain'),
   kReset: Symbol('reset'),
   kDestroyed: Symbol.for('nodejs.stream.destroyed'),
+  kResume: Symbol('resume'),
+  kOnError: Symbol('on error'),
   kMaxHeadersSize: Symbol('max headers size'),
   kRunningIdx: Symbol('running index'),
   kPendingIdx: Symbol('pending index'),
